refactor obj viewer
